// src/clients/flinkArtifacts/models/CreateArtifactV1FlinkArtifactRequest.ts
export interface CreateArtifactV1FlinkArtifactRequest {
    /**
     * Cloud provider where the Flink Artifact archive is uploaded.
     * @type {string}
     * @memberof CreateArtifactV1FlinkArtifactRequest
     */
    cloud: string;
    /**
     * The Cloud provider region the Flink Artifact archive is uploaded.
     * @type {string}
     * @memberof CreateArtifactV1FlinkArtifactRequest
     */
    region: string;
    /**
     * Environment the Flink Artifact belongs to.
     * @type {string}
     * @memberof CreateArtifactV1FlinkArtifactRequest
     */
    environment: string;
    /**
     * Unique name of the Flink Artifact per cloud, region, environment scope.
     * @type {string}
     * @memberof CreateArtifactV1FlinkArtifactRequest
     */
    display_name: string;
    /**
     * Java class or alias for the artifact as provided by developer. Deprecated
     * @type {string}
     * @memberof CreateArtifactV1FlinkArtifactRequest
     * @deprecated
     */
    _class?: string;
    /**
     * Archive format of the Flink Artifact.
     * @type {string}
     * @memberof CreateArtifactV1FlinkArtifactRequest
     */
    content_format?: string;
    /**
     * Description of the Flink Artifact.
     * @type {string}
     * @memberof CreateArtifactV1FlinkArtifactRequest
     */
    description?: string;
    /**
     * Documentation link of the Flink Artifact.
     * @type {string}
     * @memberof CreateArtifactV1FlinkArtifactRequest
     */
    documentation_link?: string;
    /**
     * Runtime language of the Flink Artifact.
     * @type {string}
     * @memberof CreateArtifactV1FlinkArtifactRequest
     */
    runtime_language?: string;
    /**
     * 
     * @type {CreateArtifactV1FlinkArtifactRequestUploadSource}
     * @memberof CreateArtifactV1FlinkArtifactRequest
     */
    upload_source: CreateArtifactV1FlinkArtifactRequestUploadSource;
}